buckets/external_authentication/requirements.txt
google-cloud==0.31.0
google-api-python-client==1.6.4
google-auth-httplib2==0.0.3
oauth2client==4.1.2
sendgrid